requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras pyproject.toml
#
blinker==1.9.0
    # via flask
click==8.2.1
    # via flask
flask==3.1.2
    # via
    #   prometheus-flask-exporter
    #   user (pyproject.toml)
gunicorn==23.0.0
    # via user (pyproject.toml)
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
packaging==25.0
    # via gunicorn
prometheus-client==0.22.1
    # via prometheus-flask-exporter
prometheus-flask-exporter==0.23.2
    # via user (pyproject.toml)
psycopg2-binary==2.9.10
    # via user (pyproject.toml)
werkzeug==3.1.3
    # via flask
